+++ system/src/Grav/Common/Page/Page.php
@@ -152,6 +152,23 @@ class Page
         return $this;
     }
 
+    protected function processFrontmatter()
+    {
+        // Quick check for twig output tags in frontmatter if enabled
+        if (Utils::contains($this->frontmatter, '{{')) {
+            $process_fields = $this->file()->header();
+            $ignored_fields = [];
+            foreach ((array)Grav::instance()['config']->get('system.pages.frontmatter.ignore_fields') as $field) {
+                if (isset($process_fields[$field])) {
+                    $ignored_fields[$field] = $process_fields[$field];
+                    unset($process_fields[$field]);
+                }
+            }
+            $text_header = Grav::instance()['twig']->processString(json_encode($process_fields), ['page'=>$this]);
+            $this->header((object) (json_decode($text_header, true) + $ignored_fields));
+        }
+    }
+
     /**
      * Return an array with the routes of other translated languages
      * @return array the page translated languages
@@ -290,9 +307,13 @@ class Page
                     $this->frontmatter = $file->frontmatter();
                     $this->header = (object)$file->header();
 
-                    // If there's a `frontmatter.yaml` file merge that in with the page header
-                    // note page's own frontmatter has precedence and will overwrite any defaults
                     if (!Utils::isAdminPlugin()) {
+                        // Process frontmatter with Twig if enabled
+                        if (Grav::instance()['config']->get('system.pages.frontmatter.process_twig') === true) {
+                            $this->processFrontmatter();
+                        }
+                        // If there's a `frontmatter.yaml` file merge that in with the page header
+                        // note page's own frontmatter has precedence and will overwrite any defaults
                         $frontmatter_file = $this->path . '/' . $this->folder . '/frontmatter.yaml';
                         if (file_exists($frontmatter_file)) {
                             $frontmatter_data = (array)Yaml::parse(file_get_contents($frontmatter_file));
@@ -313,6 +334,8 @@ class Page
                 }
                 $var = true;
             }
+
+
         }
 
         if ($var) {